Move select-arhive-type-screen to `glam` folder
PER-9753: Glam Onboarding Header
