+++ src/app/onboarding/components/glam/select-archive-type-screen/select-archive-type-screen.component.scss
@@ -1,5 +1,5 @@
 /* @format */
-@import '../welcome-screen/welcome-screen.component.scss';
+@import '../../welcome-screen/welcome-screen.component.scss';
 @import 'variables';
 @import 'fonts';
