+++ src/app/onboarding/onboarding.module.ts
@@ -22,7 +22,7 @@ import { WelcomeScreenComponent } from './components/welcome-screen/welcome-scre
 import { CreateNewArchiveComponent } from './components/create-new-archive/create-new-archive.component';
 import { ArchiveTypeSelectComponent } from './components/archive-type-select/archive-type-select.component';
 import { ArchiveCreationStartScreenComponent } from './components/glam/archive-creation-start-screen/archive-creation-start-screen.component';
-import { SelectArchiveTypeScreenComponent } from './components/select-archive-type-screen/select-archive-type-screen.component';
+import { SelectArchiveTypeScreenComponent } from './components/glam/select-archive-type-screen/select-archive-type-screen.component';
 import { NameArchiveScreenComponent } from './components/glam/name-archive-screen/name-archive-screen.component';
 import { GlamArchiveTypeSelectComponent } from './components/glam/archive-type-select/archive-type-select.component';
 import { ArchiveTypeIconComponent } from './components/glam/archive-type-icon/archive-type-icon.component';
